fix: gen metrics for certs with same CN diff OUs
Resolves https://github.com/aarnaud/vault-pki-exporter/issues/28

Implements logic to consider both CN and OU as unique values for a
certificate in the certificate map. This way we get metrics for certs
that have the same CN but different Subject OUs - before, only CN was
considered.

This should still handle renewals correctly for the same cert, updating
metrics in-place. That is why serial number or certificate fingerprint
was not used as certs renewed in-place should retain their time series.

I debated also adding Issuer information to uniquely identify these
certs but did not think it was needed