fix: delete revoked CRL metrics
When we duplicated the revoked certificate logic in loadCerts, it caused
a bug where revoked certificates would not have their time series
deleted because it compared certs in the map to revoked certificates to
find what to delete. And revoked certificates were never re-added to the
map.

Instead on each Prometheus refresh run, it will delete any time series
matching the exact serial number of any revoked certificates.

I figured the partial match and delete was okay when using the serial
number but it is not as strict/terse as the former deletion method
matching every label.

Resolves https://github.com/aarnaud/vault-pki-exporter/issues/34